Make the user daemon use the agents ConfigMap
Moves the bulk of `CanIntercept` to the cluster-side traffic-manager
in the form of a `PrepareIntercept` method. This method is responsible
for creating the initial agent `ConfigMap` in a namespace and to add
the `create` request into that is picked up by the watcher and results
in a generated entry.

All `Uninstall` functionality is now done using the helm chart and the
agent `ConfigMap`.

This change makes the telepresence multi-port aware, i.e. it can have
multiple simultaneous intercepts on different container ports in the
same pod.

+++ pkg/forwarder/forwarder.go
@@ -27,7 +27,7 @@ type Forwarder struct {
 	tCtx       context.Context
 	tCancel    context.CancelFunc
 	targetHost string
-	targetPort int32
+	targetPort uint16
 
 	manager     manager.ManagerClient
 	sessionInfo *manager.SessionInfo
@@ -36,7 +36,7 @@ type Forwarder struct {
 	mgrVersion semver.Version
 }
 
-func NewForwarder(listen *net.TCPAddr, targetHost string, targetPort int32) *Forwarder {
+func NewForwarder(listen *net.TCPAddr, targetHost string, targetPort uint16) *Forwarder {
 	return &Forwarder{
 		listenAddr: listen,
 		targetHost: targetHost,
@@ -114,7 +114,7 @@ func (f *Forwarder) Close() error {
 	return nil
 }
 
-func (f *Forwarder) Target() (string, int32) {
+func (f *Forwarder) Target() (string, uint16) {
 	f.mu.Lock()
 	defer f.mu.Unlock()
